pyproject.toml
[tool.poetry]
name = "evcs"
version = "0.1.0"
description = ""
authors = ["José Barroso <jose.renato77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.22"
alembic = "^1.12.1"
celery = {extras = ["redis"], version = "^5.3.4"}
psycopg2-binary = "^2.9.9"
pytest-mock = "^3.12.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
taskipy = "^1.12.0"
blue = "^0.9.1"
ruff = "^0.1.3"
httpx = "^0.25.0"
isort = "^5.12.0"
factory-boy = "^3.3.0"
pytest-mock = "^3.12.0"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
exclude = [".venv", "migrations"]

[too.isort]
profile = "black"
line_length = 79
extend_skip = ["migrations"]

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = "(migrations/)"

[tool.taskipy.tasks]
lint = "ruff . && blue --check . --diff"
format = "blue . && isort ."
run = "uvicorn evcs.app:app --reload"
pre_test = "task lint"
test = "pytest -s -x --cov=evcs -vv"
post_test = "coverage html"
